fix(task): harden runner and test boundaries
